+++ test/write/two.sv
@@ -1,3 +1,9 @@
+package P;
+    typedef logic T;
+endpackage
+module three;
+    P::T x;
+endmodule
 module two;
     logic x;
 endmodule